requirements.txt
# Core libraries
fastapi
async-timeout

# Database
SQLAlchemy
asyncpg
aiopg

# Migration
alembic
psycopg2-binary

# Asynchronous
httpx
aiohttp
aiosignal
amqp
anyio
asyncio

# Data manipulation
pydantic
pydantic[email]
pandas

#Utils
PyJWT==2.4.0
passlib==1.7.4
validators
python-dotenv
bcrypt

# Testing
pytest
pytest-asyncio
pytest-metadata
pytest-xdist
pytest-html
faker
openpyxl==3.0.10

# HTTP
uvicorn
gunicorn

# Additional dependencies
# Use the git repository for the latest version of a package
# git+https://github.com/vzip/wallet.git